fixed bugs

+++ Slower Than Light/src/logic/processors/TimeHolder.java
@@ -21,7 +21,7 @@ public class TimeHolder extends TimerTask{
      * Creates an object of the type Timeholder
      * @param gameTime The amount of time in seconds which the game takes
      */
-    public TimeHolder(int gameTime, int oxygenTime)
+    public TimeHolder(double gameTime, double oxygenTime)
     {
         saboteurCountdown = 5;
         timeLeft = gameTime;
